functions/highlight-detection/main.py
import functions_framework
from google.cloud import firestore
import json
import requests
from vertexai.language_models import TextGenerationModel
import vertexai
import os

# Initialize clients
db = firestore.Client()
PROJECT_ID = os.getenv('GOOGLE_CLOUD_PROJECT')

@functions_framework.http
def process_play(request):
    """Cloud Function entry point"""
    try:
        request_json = request.get_json()
        
        if not request_json:
            return 'No data provided', 400
        
        game_pk = request_json.get('game_pk')
        play_id = request_json.get('play_id')
        
        if not game_pk or not play_id:
            return 'Missing required parameters', 400
            
        # Initialize Vertex AI
        vertexai.init(project=PROJECT_ID, location="us-central1")
        
        return json.dumps({
            'success': True,
            'message': 'Processed successfully'
        })
        
    except Exception as e:
        return f'Error: {str(e)}', 500