functions/highlight-detection/requirements.txt
functions-framework==3.*
google-cloud-aiplatform==1.*
google-cloud-firestore==2.*
vertexai==1.*
requests==2.*